Very major fix for the embedded Woopra Analytics
git-svn-id: https://plugins.svn.wordpress.org/woopra/trunk@67343 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ woopra_xml.php
@@ -33,7 +33,7 @@ class WoopraAPI {
 		{
 			return false;	//	they both need to be set for this to work
 		}
-		$this->server = floor(((int)($this->siteid))/100000000);
+		$this->server = floor(((float)($this->siteid))/100000000); 
 		return true;
 	}
